Remove RadixEngineAtomProcessor
