+++ radixdlt/src/main/java/com/radixdlt/consensus/tempo/ChainedBFT.java
@@ -18,9 +18,22 @@
 package com.radixdlt.consensus.tempo;
 
 import com.google.inject.Inject;
+import com.radixdlt.common.AID;
+import com.radixdlt.common.Atom;
 import com.radixdlt.consensus.Consensus;
 import com.radixdlt.consensus.ConsensusObservation;
+import com.radixdlt.constraintmachine.CMError;
+import com.radixdlt.constraintmachine.DataPointer;
+import com.radixdlt.constraintmachine.Particle;
+import com.radixdlt.engine.AtomEventListener;
+import com.radixdlt.engine.RadixEngine;
+import com.radixdlt.middleware2.converters.AtomToBinaryConverter;
 import com.radixdlt.store.LedgerEntry;
+import com.radixdlt.store.LedgerEntryStore;
+import com.radixdlt.universe.Universe;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
 import org.radix.logging.Logger;
 import org.radix.logging.Logging;
 import org.radix.utils.SimpleThreadPool;
@@ -39,20 +52,46 @@ public final class ChainedBFT implements Consensus, Closeable {
 
 	private final BlockingQueue<ConsensusObservation> consensusObservations;
 	private final SimpleThreadPool<LedgerEntry> consensusThreadPool;
+	private final EventCoordinator eventCoordinator;
+	private final AtomToBinaryConverter atomToBinaryConverter;
+	private final Universe universe;
+	private final LedgerEntryStore store;
+	private final RadixEngine radixEngine;
 
 	@Inject
-	public ChainedBFT(MemPool memPool) {
+	public ChainedBFT(
+		EventCoordinator eventCoordinator,
+		MemPool memPool,
+		AtomToBinaryConverter atomToBinaryConverter,
+		Universe universe,
+		RadixEngine radixEngine,
+		LedgerEntryStore store
+	) {
 		Objects.requireNonNull(memPool);
+		Objects.requireNonNull(eventCoordinator);
+		Objects.requireNonNull(atomToBinaryConverter);
+		Objects.requireNonNull(universe);
+		Objects.requireNonNull(store);
+		Objects.requireNonNull(radixEngine);
+
+		this.eventCoordinator = eventCoordinator;
+		this.atomToBinaryConverter = atomToBinaryConverter;
+		this.universe = universe;
+		this.store = store;
+		this.radixEngine = radixEngine;
 
 		this.consensusObservations = new LinkedBlockingQueue<>(INBOUND_QUEUE_CAPACITY);
 
 		this.consensusThreadPool = new SimpleThreadPool<>("Consensus", 1, memPool::takeNextEntry, this::doConsensus, log);
 		this.consensusThreadPool.start();
+		this.initGenesis();
 	}
 
 	private void doConsensus(LedgerEntry entry) {
 		// stupid simple "consensus", just immediately commit anything we get our hands on
 		this.consensusObservations.add(ConsensusObservation.commit(entry));
+		Atom atom = atomToBinaryConverter.toAtom(entry.getContent());
+		eventCoordinator.processProposal(atom);
 	}
 
 	@Override
@@ -63,5 +102,61 @@ public final class ChainedBFT implements Consensus, Closeable {
 	@Override
 	public void close() {
 		this.consensusThreadPool.stop();
+	}
+
+	private void initGenesis() {
+		try {
+			LinkedList<AID> atomIds = new LinkedList<>();
+			for (Atom atom : universe.getGenesis()) {
+				if (!store.contains(atom.getAID())) {
+					radixEngine.store(atom,
+						new AtomEventListener() {
+							@Override
+							public void onCMSuccess(Atom atom) {
+								log.debug("Genesis Atom " + atom.getAID() + " stored to atom store");
+							}
+
+							@Override
+							public void onCMError(Atom atom, CMError error) {
+								log.fatal("Failed to addAtom genesis Atom: " + error.getErrorCode() + " "
+									+ error.getErrMsg() + " " + error.getDataPointer() + "\n"
+									+ atom + "\n"
+									+ error.getCmValidationState().toString());
+								System.exit(-1);
+							}
+
+							@Override
+							public void onVirtualStateConflict(Atom atom, DataPointer dp) {
+								log.fatal("Failed to addAtom genesis Atom: Virtual State Conflict");
+								System.exit(-1);
+							}
+
+							@Override
+							public void onStateConflict(Atom atom, DataPointer dp, Atom conflictAtom) {
+								log.fatal("Failed to addAtom genesis Atom: State Conflict");
+								System.exit(-1);
+							}
+
+							@Override
+							public void onStateMissingDependency(AID atomId, Particle particle) {
+								log.fatal("Failed to addAtom genesis Atom: Missing Dependency");
+								System.exit(-1);
+							}
+						});
+				}
+			}
+			waitForAtoms(atomIds);
+		} catch (Exception ex) {
+			log.fatal("Failed to addAtom genesis Atom", ex);
+			System.exit(-1);
+		}
+	}
+
+	private void waitForAtoms(List<AID> atomHashes) throws InterruptedException {
+		for (AID atomID : atomHashes) {
+			while (!store.contains(atomID)) {
+				TimeUnit.MILLISECONDS.sleep(100);
+			}
+		}
 	}
 }
