+++ radixdlt/src/main/java/org/radix/api/http/RadixHttpServer.java
@@ -19,7 +19,6 @@ package org.radix.api.http;
 
 import com.radixdlt.consensus.tempo.MemPool;
 import com.radixdlt.middleware2.converters.AtomToBinaryConverter;
-import com.radixdlt.middleware2.processing.RadixEngineAtomProcessor;
 import com.radixdlt.serialization.DsonOutput;
 import com.radixdlt.serialization.Serialization;
 import com.radixdlt.store.LedgerEntryStore;
