+++ radixdlt/src/main/java/org/radix/api/services/InternalService.java
@@ -19,7 +19,6 @@ package org.radix.api.services;
 
 import com.radixdlt.common.Atom;
 import com.radixdlt.consensus.tempo.MemPool;
-import com.radixdlt.middleware2.processing.RadixEngineAtomProcessor;
 import com.radixdlt.serialization.Serialization;
 import com.radixdlt.universe.Universe;
 import com.radixdlt.utils.Bytes;
